+++ packages/builder/plugin-swc/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/builder-plugin-swc
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/builder-shared@2.42.2
+- @modern-js/utils@2.42.2
+
 ## 2.42.1
 
 ### Patch Changes