+++ packages/cli/plugin-data-loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-data-loader
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.42.2
+- @modern-js/utils@2.42.2
+
 ## 2.42.1
 
 ### Patch Changes