+++ packages/server/plugin-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-server
 
+## 2.42.2
+
+### Patch Changes
+
+- Updated dependencies [07c56c0]
+  - @modern-js/server-utils@2.42.2
+  - @modern-js/utils@2.42.2
+
 ## 2.42.1
 
 ### Patch Changes