+++ packages/toolkit/compiler/babel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/babel-compiler
 
+## 2.42.2
+
+### Patch Changes
+
+- @modern-js/utils@2.42.2
+
 ## 2.42.1
 
 ### Patch Changes